chore: clean up posthog analytics
